Fix SPIRV generation bug caused by swizzled types.
Swizzle FullTypes are returned with a swizzle type, but the Type* will always be the final (swizzled) type. The SPIRV generator needs to create a right hand vector for assigning a value to a swizzle mask, to then perform a shuffle, which means it needs to extract the real vector type of the left hand side.