Fix diagnostics and optimize lexing
Lexing is done by creating a category of each allowed character.
Instead of doing lots of comparisons with > and < for ranges, we lookup a bit to know if the character belongs to a class.
This can be SIMD:ified for even faster lexing.
Multi character operators are now concatenated before matched against allowed words, reducing the amount of lookups.