Add transient shader parameter modifiers
Modifiers like no_perspective and no_interpolation are only available in the pixel shader. But for correct linkage, the previous geometry shader must provide the same qualifier as output, which is standard in glsl and hlsl. However, since we compose our programs in code, we can just infer that the qualifier can be transferred backwards to the previous shader stage.