Lots of various improvements.
It's now possible to get a pointer to uniform data in a scope code, and pass it to a function. Let's say you'd want to get a pointer to a member of a struct which is bound as a uniform resource, now you can.
Atomics are separated into workgroup and uniform atomics as overloads. This means atomic operations on local variables are not allowed, but mutable uniforms can use atomic operations, and workgroup variables can too.