Various small fixes
Produce parse error when a keyword is placed incorrectly. We have so many keywords that accidentally using an identifier with that name is easy, so to avoid this issue, the compiler should emit errors when a possible identifier was parsed as a keyword.
Fixed some header generation issues.
Added a deserialized size variable in the binary, so that we can avoid newing every single shader object on the client side.
Changed physical buffer address SPIRV extension to use the KHR one, instead of EXT.