Various fixes
Adds an extra uint at the end of the reflection binary to allow for the runtime to understand how much memory is required for the deserialized version of the reflection data. This allows the runtime to make a single allocation for it.
Added sample count, sample shading and alpha-to-coverage and alpha-to-one for render states.
Fixed bug where duplicate programs wouldn't get picked up.